requirements.txt
openai>=1.44.0
requests>=2.31.0
ddgs>=9.5.4
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
lxml>=4.9.3
python-pptx>=0.6.23
tenacity>=8.2.3